requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-default-group-wheel==1.2.2
comm==0.1.4
dataclasses-json==0.5.14
datasets==2.14.5
debugpy==1.6.7.post1
decorator==5.1.1
dill==0.3.7
diskcache==5.6.1
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
ijson==3.2.3
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.273
langsmith==0.0.26
linkedin-scraper==2.11.2
llama-cpp-python==0.1.79
llm-llama-cpp==0.1a0
lxml==4.9.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numexpr==2.8.5
numpy==1.25.2
openai==0.27.9
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
playwright==1.37.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pyee==9.0.4
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ulid==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlite-fts4==1.0.3
sqlite-utils==3.35
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.2
